pyproject.toml
[project]
name = "ai-network-advisor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.65.0",
    "python-decouple>=3.8",
]
